+++ bflibrary/include/bfiff.h
@@ -29,7 +29,7 @@ extern "C" {
 #pragma pack(1)
 
 struct TbIffPalette { // sizeof=4
-    long Loaded; // offset=0
+    s32 Loaded; // offset=0
     ubyte Palette[256]; // offset=4
 };
 
@@ -38,7 +38,7 @@ typedef struct TbIffPalette TbIffPalette;
 struct TbIff { // sizeof=18
     ushort Width; // offset=0
     ushort Height; // offset=2
-    ulong FileSize; // offset=4
+    u32 FileSize; // offset=4
     TbIffPalette *Palette; // offset=8
     ushort XOffset; // offset=12
     ushort YOffset; // offset=14
@@ -74,7 +74,7 @@ TbResult LbIffLoad(const char *fname, ubyte *buf, struct TbIff *iff);
  * @return
  */
 TbResult LbIffSave(const char *fname, const TbPixel *inp_buffer,
-  ulong width, ulong height, const ubyte *pal, TbBool force_fname);
+  u32 width, u32 height, const ubyte *pal, TbBool force_fname);
 
 /** Save "ILBM" IFF Interleaved Bitmap from screen sized buffer.
  *
